Update cnc-ddraw.yml
archive_extract returns an error code 2 with most zip provided by FunkyFr3sh/cnc-ddraw. Seems to be related to 'cnc-ddraw config.exe' being extracted as a directory and crashing 7z. A solution would be to append an argument to extract only a selection of files and avoid the one crashing, but I cannot figure out this action. As an alternative, cnc-ddraw-experimental-debuglog.zip does not contain the problematic exe and the extraction is successful, although without Shaders. The priority is still to provide a working dll for software requiring it, while Shaders are not necessary.